Refactor footer into its own folder structure
