+++ client/src/components/Footer/Footer.jsx
@@ -1,5 +1,3 @@
-import React from "react";
-
 const Footer = () => {
   return (
     <footer className="w-full h-[40vh] bg-gray-800 text-white flex items-center justify-center">
