+++ client/src/layout/HomeLayout.jsx
@@ -1,7 +1,7 @@
 import Header from "../components/Header";
 import Animals from "../components/Animals";
 import Component2 from "../components/Component2";
-import Footer from "../components/Footer";
+import Footer from "../components/Footer/Footer";
 
 const HomeLayout = () => {
   return (
