+++ client/src/layout/MapLayout.jsx
@@ -1,6 +1,6 @@
 import Header from "../components/Header";
 import Component2 from "../components/Component2";
-import Footer from "../components/Footer";
+import Footer from "../components/Footer/Footer";
 
 const MapLayout = () => {
   return (
